pyproject.toml
[project]
name = "imply-druid-mcp"
version = "0.1.1"
description = "MCP server for Imply Cloud/Druid - query and manage data through AI assistants"
requires-python = ">=3.10"
readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "YB Hwang" }
]
keywords = [
    "mcp",
    "model-context-protocol",
    "imply",
    "druid",
    "apache-druid",
    "database",
    "ai",
    "llm",
    "claude"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Database",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/yeongbin-hwang/imply-druid-mcp"
Repository = "https://github.com/yeongbin-hwang/imply-druid-mcp"
Issues = "https://github.com/yeongbin-hwang/imply-druid-mcp/issues"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "black>=24.0.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
imply-druid-mcp = "imply_druid_mcp.server:run"

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
